# jualdanbeli-STANDALONE/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: jualdanbeli Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ──────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /messages/conversations:
    get:
      operationId: getConversations
      tags: [messages]
      summary: List all conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

    post:
      operationId: startConversation
      tags: [messages]
      summary: Start a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /messages/conversations/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/conversations/{conversationId}/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── NOTIFICATIONS ──────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ─── AUTH ──────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password, phone]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [buyer, seller]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── USERS ─────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        sellerInfo:
          $ref: "#/components/schemas/SellerInfo"
    UserProfile:
      type: object
      required: [id, name, role, createdAt, stats]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        stats:
          $ref: "#/components/schemas/UserStats"
        sellerInfo:
          $ref: "#/components/schemas/SellerInfo"
    UserStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalSales:
          type: integer
        averageRating:
          type: number
        totalReviews:
          type: integer
      required: [totalProducts, totalSales, averageRating, totalReviews]
    SellerInfo:
      type: object
      properties:
        shopName:
          type: ["string", "null"]
        shopDescription:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        isVerified:
          type: boolean
      required: [isVerified]
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        iconUrl:
          type: ["string", "null"]
        productCount:
          type: integer

    # ─── PRODUCTS ──────────────────────────────────────────────────
    Product:
      type: object
      required: [id, name, description, price, stock, condition, status, sellerId, categoryId, images, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        stock:
          type: integer
        condition:
          type: string
          enum: [new, used]
        status:
          type: string
          enum: [active, pending, flagged, removed]
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/UserProfile"
        categoryId:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        images:
          type: array
          items:
            type: string
        weight:
          type: ["number", "null"]
        city:
          type: ["string", "null"]
        averageRating:
          type: number
        totalReviews:
          type: integer
        totalSold:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Cart:
      type: object
      required: [id, userId, items, subtotal]
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
    CartItem:
      type: object
      required: [id, productId, quantity, unitPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        unitPrice:
          type: number
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # ─── ORDERS ────────────────────────────────────────────────────
...
    Conversation:
      type: object
      required: [id, participants, lastMessage, updatedAt]
      properties:
        id:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        productId:
          type: ["integer", "null"]
        product:
          $ref: "#/components/schemas/Product"
        lastMessage:
          type: ["string", "null"]
        unreadCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [recipientId]
      properties:
        recipientId:
          type: integer
        productId:
          type: integer
        initialMessage:
          type: string
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ─── NOTIFICATIONS ─────────────────────────────────────────────
...